Time modification implementation
Implemented more buttons for time modification: can add/remove minutes
and 5 seconds. Also added a manual choice for adding minutes and
seconds to the timer.

If time duration modifiers are used during the timeout the time that is
modified is that of the underlying state, not the timeout.

Bugs:
Noticed a few bugs regarding the timeout timer if the pause button is
used during the timeout. Also need to check the reaction of the “cancel
mistake” button if used during a timeout.